+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-25T14:18:53.694Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-26T09:10:19.431Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")