+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-08-25T14:18:57.036Z")
+    created: Optional[datetime] = parse_datetime("2025-08-26T09:10:22.707Z")
     r"""The date and time the resource was created"""