+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K3GSHHH1H8996XC3GNTW6C78"
+        "01K3JT97W6EZ21Y5VE1V3M9G61"
     )
     r"""The id of the resource"""