+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -198,5 +198,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-25T14:18:53.694Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-26T09:10:19.431Z")
     r"""The date and time the resource was last updated"""