+++ packages/orq-rc/RELEASES.md
@@ -2359,3 +2359,13 @@ Based on:
 - [python v3.12.0-rc.12] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.12] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.12 - packages/orq-rc
+
+## 2025-08-26 09:11:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.5 (2.687.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.13] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.13] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.13 - packages/orq-rc